Move test dir creation macro to new lib crate to be shared
