+++ Cargo.toml
@@ -10,6 +10,7 @@ edition = "2018"
 members = [
     "xtask",
     "tests/print_file_to_stdout",
+    "pycors_test_helpers",
 ]
 
 [dependencies]
@@ -53,6 +54,7 @@ mockall = "0.6"
 assert_cmd = "0.12"
 predicates = "1"
 indoc = "0.3"
+pycors_test_helpers = {path = "pycors_test_helpers"}
 # pretty_assertions = "0.6"
 
 [target.'cfg(not(target_os = "windows"))'.dev-dependencies.users]
