+++ pycors_test_helpers/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "pycors_test_helpers"
+version = "0.1.0"
+authors = ["Nicolas Bigaouette <nbigaouette@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+env_logger = "0.7"
