+++ src/toolchain/tests.rs
@@ -13,110 +13,9 @@ use std::{
     process::{ExitStatus, Output},
 };
 
+use pycors_test_helpers::{create_test_temp_dir, mock_executable, MockedOutput};
+
 use crate::{constants::INFO_FILE, utils::directory::MockPycorsHomeProviderTrait};
-
-#[cfg(windows)]
-const EXEC_EXTENSION: &str = ".exe";
-#[cfg(not(windows))]
-const EXEC_EXTENSION: &str = "";
-
-struct MockedOutput<'a> {
-    out: Option<&'a str>,
-    err: Option<&'a str>,
-}
-
-fn mock_executable<P, S>(
-    executable_location: P,
-    executable_name: S,
-    output: MockedOutput,
-) -> Result<()>
-where
-    P: AsRef<Path>,
-    S: AsRef<str>,
-{
-    _mock_executable(
-        executable_location.as_ref(),
-        executable_name.as_ref(),
-        output,
-    )
-}
-
-fn _mock_executable(
-    executable_location: &Path,
-    executable_name: &str,
-    output: MockedOutput,
-) -> Result<()> {
-    let _cargo_output = std::process::Command::new("cargo")
-        .args(&["build", "--package", "print_file_to_stdout"])
-        .output()
-        .with_context(|| "Failed to execute 'cargo build --package print_file_to_stdout")?;
-
-    if !executable_location.exists() {
-        fs::create_dir_all(&executable_location)?
-    }
-
-    let stdout_filepath = executable_location.join(format!(
-        "{}{}_pycors_tests_to_print_stdout.txt",
-        executable_name, EXEC_EXTENSION
-    ));
-    let stderr_filepath = executable_location.join(format!(
-        "{}{}_pycors_tests_to_print_stderr.txt",
-        executable_name, EXEC_EXTENSION
-    ));
-
-    if stdout_filepath.exists() {
-        fs::remove_file(&stdout_filepath)
-            .with_context(|| format!("Failed to remove file {:?}", stdout_filepath))?;
-    }
-    if stderr_filepath.exists() {
-        fs::remove_file(&stderr_filepath)
-            .with_context(|| format!("Failed to remove file {:?}", stderr_filepath))?;
-    }
-
-    if let Some(stdout) = output.out {
-        let mut f = File::create(&stdout_filepath)
-            .with_context(|| format!("Failed to create file {:?}", stdout_filepath))?;
-        f.write_all(stdout.as_bytes())
-            .with_context(|| format!("Failed to write to file {:?}", stdout_filepath))?;
-    }
-    if let Some(stderr) = output.err {
-        let mut f = File::create(&stderr_filepath)
-            .with_context(|| format!("Failed to create file {:?}", stderr_filepath))?;
-        f.write_all(stderr.as_bytes())
-            .with_context(|| format!("Failed to write to file {:?}", stderr_filepath))?;
-    }
-
-    let print_file_to_stdout = {
-        let target_dir = match env::var("CARGO_TARGET_DIR") {
-            Ok(dir) => dir,
-            Err(_) => String::from("target"),
-        };
-
-        #[cfg_attr(not(windows), allow(unused_mut))]
-        let mut tmp = Path::new(&target_dir)
-            .join("debug")
-            .join("print_file_to_stdout");
-
-        #[cfg(windows)]
-        tmp.set_extension("exe");
-
-        tmp
-    };
-
-    fs::copy(
-        &print_file_to_stdout,
-        executable_location.join(format!("{}{}", executable_name, EXEC_EXTENSION)),
-    )
-    .with_context(|| {
-        format!(
-            "Failed to copy {:?} to {:?}",
-            print_file_to_stdout,
-            executable_location.join(format!("{}{}", executable_name, EXEC_EXTENSION))
-        )
-    })?;
-
-    Ok(())
-}
 
 #[test]
 fn version_or_path_from_str_success_major_minor_patch() {
