+++ tests/integration/mod.rs
@@ -9,45 +9,11 @@ use indoc::indoc;
 use predicates::prelude::*;
 
 use pycors::{
-    constants::{home_env_variable, EXECUTABLE_NAME, INFO_FILE, TOOLCHAIN_FILE},
+    constants::{home_env_variable, INFO_FILE, TOOLCHAIN_FILE},
     Result,
 };
 
-// https://stackoverflow.com/a/40234666
-macro_rules! function_path {
-    () => {{
-        fn f() {}
-        fn type_name_of<T>(_: T) -> &'static str {
-            std::any::type_name::<T>()
-        }
-        let name = type_name_of(f);
-        &name[..name.len() - 3]
-    }};
-}
-
-macro_rules! create_test_temp_dir {
-    () => {{
-        let dir = env::temp_dir()
-            .join(EXECUTABLE_NAME)
-            .join("integration_tests");
-
-        if !dir.exists() {
-            fs::create_dir_all(&dir).unwrap();
-        }
-        let mut dir = dir.canonicalize().unwrap();
-        for component in function_path!().split("::").skip(1) {
-            dir.push(component);
-        }
-
-        if dir.exists() {
-            fs::remove_dir_all(&dir).unwrap();
-        }
-
-        fs::create_dir_all(&dir).unwrap();
-
-        dir
-    }};
-}
+use pycors_test_helpers::{create_test_temp_dir, mock_executable, MockedOutput};
 
 mod help;
 mod install;
